% Generated manually
\docType{package}
\name{cdmAcm0019-package}
\alias{cdmAcm0019}
\alias{cdmAcm0019-package}
\title{cdmAcm0019: Implementation of CDM Methodology ACM0019}
\description{Tidyverse-aligned helpers implementing the Clean Development Mechanism methodology ACM0019 for nitrous oxide abatement in nitric acid plants.}
\seealso{\url{https://cdm.unfccc.int/methodologies/DB/HKCO7RKOQO748WNXJNDEW3BJT9XN8L}}
\author{OpenAI Assistant <assistant@independentimpact.org>}
